refactor: rewrite in Nuxt (#62)
It's been quite a while since I've last updated my site, and I'd love to
give it a refresh at some point in the near future. This is a rewrite of
the site in Nuxt to enable a component-based structure, deployment
targets other than GitHub Pages, and a new foundation for my little
corner of the web.